requirements.txt
# Core Dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
numpy>=1.24.0

# Database Operations
sqlalchemy>=2.0.0

# Web Development
flask>=3.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Machine Learning
scikit-learn>=1.3.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Security
cryptography>=41.0.0

# Performance
memory-profiler>=0.61.0

# Testing
pytest>=7.4.0
pytest-mock>=3.12.0

# DevOps Tools (optional)
boto3>=1.28.0
azure-identity>=1.14.0
azure-mgmt>=0.20.0
kubernetes>=28.1.0